Rewrite the commit action
We don't need to resort to including the class and calling it from PHP.
There's inherent risk of that interface changing in future releases.

The binary for PHPCS itself has a sufficient interface to differentiate
between error and non-error condition.

Future improvement would be to manually parse the configuration xml,
and then map the changed files to the configured directories, so we only
test what is changed.